# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Scandere AI Store API",
        "description": "REST API for digital products e-commerce platform. Sell and manage digital documents (PDFs, Excel, Word) with secure downloads, payments via Stripe, and comprehensive admin panel.",
        "contact": {
            "name": "Scandere Support",
            "email": "team@scandere.info"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.scandereai.store",
            "description": "Production API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "View shopping cart",
                "description": "Get cart items with automatic bundle detection and savings calculation",
                "operationId": "0652f94dde92e36bbb3c461356ce1d7b",
                "responses": {
                    "200": {
                        "description": "Cart contents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "product": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "title": {
                                                                "type": "string"
                                                            },
                                                            "price": {
                                                                "type": "number"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "subtotal": {
                                            "description": "Total price (with bundle discount applied if eligible)",
                                            "type": "number",
                                            "example": 24.99
                                        },
                                        "bundle": {
                                            "description": "Bundle applied (if cart items match a bundle)",
                                            "type": "object",
                                            "nullable": true
                                        },
                                        "bundle_savings": {
                                            "description": "Amount saved with bundle discount",
                                            "type": "number",
                                            "example": 7.98
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Clear cart",
                "description": "Remove all items from shopping cart",
                "operationId": "c11937b9c6a767730b1dc083d3d13e96",
                "responses": {
                    "200": {
                        "description": "Cart cleared",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cart cleared."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add product to cart",
                "description": "Add a product to shopping cart (prevents adding already purchased products)",
                "operationId": "8a051e160148a408fe92e1f8da586aab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "Product ID to add",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product added to cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Added to cart."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "409": {
                        "description": "Product already purchased",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Already purchased."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/{product_id}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Remove product from cart",
                "description": "Remove a specific product from shopping cart",
                "operationId": "16c3558b2c2aa39ed24b6e1c4a607c80",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product ID to remove",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product removed from cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Removed from cart."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/bundle/{bundle_id}": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add bundle to cart",
                "description": "Add all products from a bundle to cart (skips already purchased products)",
                "operationId": "24dffeedabf5b4780aed45323be36f8b",
                "parameters": [
                    {
                        "name": "bundle_id",
                        "in": "path",
                        "description": "Bundle ID to add",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bundle added to cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bundle added to cart."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Bundle not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format: Bearer {token}"
            }
        }
    },
...
}